[Refactor] Todo pakage 구조 변경
